moved some funcs to helper file, updatePackage checks for good metadata & if it exists already
